# api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CS2KZ API",
    "description": "Source Code available on [GitHub](https://github.com/KZGlobalTeam/cs2kz-api).",
    "contact": {
      "name": "AlphaKeks",
      "email": "alphakeks@dawn.sh"
    },
    "license": {
      "name": "Licensed under the GPLv3",
      "url": "https://www.gnu.org/licenses/gpl-3.0"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/bans/{ban_id}": {
      "get": {
        "tags": [
          "Bans"
        ],
        "operationId": "get",
        "parameters": [
          {
            "name": "ban_id",
            "in": "path",
            "description": "The ban's ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "uint64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "A player ban.",
                  "required": [
                    "id",
                    "player",
                    "reason",
                    "created_on"
                  ],
                  "properties": {
                    "id": {
                      "type": "integer",
                      "format": "uint64",
                      "description": "The ban's ID.",
                      "minimum": 0
                    },
                    "player": {
                      "$ref": "#/components/schemas/Player"
                    },
                    "server": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/ServerInfo"
                        }
                      ],
                      "nullable": true
                    },
                    "reason": {
                      "$ref": "#/components/schemas/BanReason"
                    },
                    "admin": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Player"
                        }
                      ],
                      "nullable": true
                    },
                    "created_on": {
                      "type": "string",
                      "format": "date-time",
                      "description": "When this ban was submitted."
                    },
                    "expires_on": {
                      "type": "string",
                      "format": "date-time",
                      "description": "When this ban will expire.",
                      "nullable": true
                    },
                    "unban": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Unban"
                        }
                      ],
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      },
      "patch": {
        "tags": [
          "Bans"
        ],
        "operationId": "patch",
        "parameters": [
          {
            "name": "ban_id",
            "in": "path",
            "description": "The ban's ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "uint64",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "422": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "security": [
          {
            "Browser Session": [
              "bans"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Bans"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "ban_id",
            "in": "path",
            "description": "The ban's ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "uint64",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUnban"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "A newly reverted ban.",
                  "required": [
                    "unban_id"
                  ],
                  "properties": {
                    "unban_id": {
                      "type": "integer",
                      "format": "uint64",
                      "description": "The unban's ID.",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "security": [
          {
            "Browser Session": [
              "bans"
            ]
          }
        ]
      }
    },
...
    "/auth/login": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "login",
        "parameters": [
          {
            "name": "redirect_to",
            "in": "query",
            "description": "Where the user wants to be redirected to after the login process is done.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            }
          }
        ],
        "responses": {
          "303": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "logout",
        "responses": {
          "303": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "security": [
          {
            "Browser Session": []
          }
        ]
      }
    },
    "/auth/callback": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "callback",
        "parameters": [
          {
            "name": "openid.ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "openid.identity",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "openid.claimed_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            }
          },
          {
            "name": "openid.mode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "openid.return_to",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            }
          },
          {
            "name": "openid.op_endpoint",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "openid.response_nonce",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "openid.invalidate_handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "openid.assoc_handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "openid.signed",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "openid.sig",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "default": null,
                  "nullable": true
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      }
    },
...
    "/admins/{steam_id}": {
      "get": {
        "tags": [
          "Admins"
        ],
        "operationId": "get",
        "parameters": [
          {
            "name": "steam_id",
            "in": "path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "title": "Steam ID",
                  "example": "STEAM_1:1:161178172"
                },
                {
                  "type": "string",
                  "title": "Steam ID3",
                  "example": "U:1:322356345"
                },
                {
                  "type": "integer",
                  "title": "Steam ID32",
                  "example": 322356345
                },
                {
                  "type": "integer",
                  "title": "Steam ID64",
                  "example": 76561198282622073
                }
              ],
              "description": "a player's SteamID",
              "example": "STEAM_1:1:161178172"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "A player with special privileges.",
                  "required": [
                    "name",
                    "steam_id",
                    "roles"
                  ],
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "The admin's name."
                    },
                    "steam_id": {
                      "$ref": "#/components/schemas/SteamID"
                    },
                    "roles": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "The admin's roles.",
                      "example": [
                        "bans",
                        "servers"
                      ]
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Admins"
        ],
        "operationId": "put",
        "parameters": [
          {
            "name": "steam_id",
            "in": "path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "title": "Steam ID",
                  "example": "STEAM_1:1:161178172"
                },
                {
                  "type": "string",
                  "title": "Steam ID3",
                  "example": "U:1:322356345"
                },
                {
                  "type": "integer",
                  "title": "Steam ID32",
                  "example": 322356345
                },
                {
                  "type": "integer",
                  "title": "Steam ID64",
                  "example": 76561198282622073
                }
              ],
              "description": "a player's SteamID",
              "example": "STEAM_1:1:161178172"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "422": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "security": [
          {
            "Browser Session": [
              "admins"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminUpdate": {
        "type": "object",
        "description": "Request body for updating admins.",
        "required": [
          "roles"
        ],
        "properties": {
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "New roles for the admin.",
            "example": [
              "bans",
              "servers"
            ]
          }
        }
      },
...
      "BanReason": {
        "type": "string",
        "description": "The different reasons for which players can be banned.",
        "enum": [
          "auto_strafe",
          "auto_bhop"
        ]
      },
      "BanUpdate": {
        "type": "object",
        "description": "Request body for updating bans.",
        "properties": {
          "reason": {
            "allOf": [
              {
                "$ref": "#/components/schemas/BanReason"
              }
            ],
            "nullable": true
          },
          "expires_on": {
            "type": "string",
            "format": "date-time",
            "description": "A new expiration date.\n\nNot specifying this at all means the expiration date will not be modified.\nIf this is explicitly `null`, the expiration date will be deleted and the ban counts as\npermanent.",
            "nullable": true
          }
        }
      },
...
      "NewUnban": {
        "type": "object",
        "description": "Request body for reverting a ban.",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string",
            "description": "The reason this ban should be reverted."
          }
        }
      },
...
      "Player": {
        "type": "object",
        "description": "A KZ player.",
        "required": [
          "name",
          "steam_id"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The player's name."
          },
          "steam_id": {
            "$ref": "#/components/schemas/SteamID"
          }
        }
      },
...
      "ServerInfo": {
        "type": "object",
        "description": "Information about a server.",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "uint16",
            "description": "The server's ID.",
            "minimum": 0
          },
          "name": {
            "type": "string",
            "description": "The server's name."
          }
        }
      },
...
      "SteamID": {
        "anyOf": [
          {
            "type": "string",
            "title": "Steam ID",
            "example": "STEAM_1:1:161178172"
          },
          {
            "type": "string",
            "title": "Steam ID3",
            "example": "U:1:322356345"
          },
          {
            "type": "integer",
            "title": "Steam ID32",
            "example": 322356345
          },
          {
            "type": "integer",
            "title": "Steam ID64",
            "example": 76561198282622073
          }
        ],
        "description": "a player's SteamID",
        "example": "STEAM_1:1:161178172"
      },
...
      "Unban": {
        "type": "object",
        "description": "A reverted ban.",
        "required": [
          "id",
          "reason",
          "created_on"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "uint64",
            "description": "The unban's ID.",
            "minimum": 0
          },
          "reason": {
            "type": "string",
            "description": "The reason for the unban."
          },
          "admin": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Player"
              }
            ],
            "nullable": true
          },
          "created_on": {
            "type": "string",
            "format": "date-time",
            "description": "When the ban was reverted."
          }
        }
      }
    },
    "securitySchemes": {
      "API Key": {
        "type": "http",
        "scheme": "bearer"
      },
      "Browser Session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "kz-auth"
      },
      "CS2 Server": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}